pulumi/main/lambda/requirements.txt
protobuf==4.21.12
google-cloud-bigquery==3.4.1
